backend/requirements.txt
requests==2.31.0
Django==5.1.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.9.0
django-filter==24.2
google-generativeai==0.3.2

